Remove unnecessary steps from Ubuntu 24.04 end-to-end job (#7369)
A recent PR converted the Ubuntu 24.04 x64 end-to-end test job to use an ephemeral 1ES-hosted agent instead of a stateful custom agent. With this conversion, we should have removed steps from the job that are only needed for stateful agents, but it was missed. This update removes the steps.

I ran the end-to-end test pipeline against this PR and confirmed that the steps are removed and the job still passes.

## Azure IoT Edge PR checklist: