Remove just-in-time install of code coverage report generator tool (#7386)
We no longer need to install the code coverage report generator tool in the dotnet checkin pipeline because it is pre-installed on the agent. Additionally, we recently started seeing blocking errors in the dotnet checkin pipeline because it tries to install the tool through the repo's Azure feed but the tool is not cached in the feed. This change removes the install commands.

## Azure IoT Edge PR checklist: