Remove 'syntax' parser directive from all Dockerfiles (#7392)
Two and a half years ago (see #6352), we added a `syntax=...` line to all our Dockerfiles to take advantage of the then-new embedded metadata (specifically, provenance information) in our images. But that change also pinned our builds to a specific version of the Dockerfile frontend...a version that is now two and a half years old.

This change removes the syntax lines from all our Dockerfiles. We'll still get the needed behaviors because they've been stable for a long time. The difference is that we'll go back to using the latest stable version of the Dockerfile frontend, which is bundled into BuildKit. This will prevent `docker buildx` from doing extra `docker pull` calls behind the scenes, and should further reduce occurances of docker.io throttling.

To test, I ran the CI Build pipeline against these changes, then ran the end-to-end test pipeline against the built images to confirm they still work. I also checked the provenance metadata on the builds manually.

## Azure IoT Edge PR checklist: