Avoid container build by adding /.mariner-toolkit-ignore-dockerenv (#7409)
Our Azure Linux builds started failing in the past week or so because Azure Linux recently made a change to how they detect container builds (see https://github.com/microsoft/azurelinux/pull/11135), so the trick we employed to force a regular build (even though we're building in a container) stopped working. This change removes the trick and takes the recommended approach of adding `/.mariner-toolkit-ignore-dockerenv` to the container.

Other changes were also needed to resolve problems that came up while testing this PR:
- Snap builds began failing, seemingly due to a bad version combination of snapcraft and lxd. I updated the build task to do `snap refresh` before building.
- Our CI build for Azure Linux on arm64 uses the build scripts from the iot-identity-service repo, so I updated this repo's references to the iot-identity-service repo to fix the container build detection problem there.
- A recent update to a dependency of cross requires a newer rustc version and was causing the edgelet CI runs to fail in this PR because our rustc version is fixed at 1.73, I added the '--locked' flag to the cargo command that installs cross in the edgelet CI build.

To test, I confirmed that the CI Build passes with these changes.

## Azure IoT Edge PR checklist: