Fix hard-coded timestamp in unit tests (#7413)
Several unit tests started failing after the new year due to a hard-coded timestamp that was supposed to represent a valid future time, but is now in the past. This change updates the test code to generate a timestamp that is one year in the future relative to DateTime.Now. I confirmed the unit tests pass once more.

## Azure IoT Edge PR checklist: