Consolidate Mariner build jobs into a single matrix (#7410)
In the packages build pipeline we have separate build jobs with a lot of duplicated YAML for building Mariner 2.0 amd64 vs. arm64. This change consolidates the separate jobs into a single matrix job. Note this also means we're now building identity service for amd64, not just arm64, for consistency (and because we'll eventually build identity service in this pipeline for all platforms, not just Mariner). This change should put us in a better position when we add support for Mariner 3.0 soon.

Additionally, this change fixes a bug: for Mariner arm64 we were already building the identity service bits but were always building them from the head of the main branch, even though the iotedge repo might be expecting bits built from an older commit. This change syncs the identity repo to the right commit first before building.

To test I ran the CI Build pipeline, confirmed it passes, and confirmed that the right artifacts are being produced. I also ran the End-to-end test pipeline using the bits from my CI build, and confirmed that the Mariner job uses the freshly-built identity service bits when it installs IoT Edge, and all tests pass.

## Azure IoT Edge PR checklist: