Build identity service in CI pipeline (#7412)
Currently, we do not build the identity service as part of the Build CI pipeline. When we need identity service packages, e.g., in most of our test pipelines, we get them from the GitHub Actions pipeline that runs in the Azure/iot-identity-service repo. We have a script to grab the packages, and it relies on a GitHub person access token (PAT) for authentication. In an effort to meet stricter security requirements that will soon be enforced, we want to eliminate PAT usage wherever possible.

This change reduces GitHub PAT usage in our test pipelines by building identity service packages directly in the Build CI pipeline and saving them as artifacts for consumption by downstream pipelines.

Also, the scripts under the platform-validation/ folder belong to a feature that was not completed and is not supported. Rather than upgrade the pipelines associated with these scripts (which also are not being used), we are removing them.

To test, I ran the Build CI pipeline against these changes and confirmed it generated all the expected artifacts. Then I ran the following pipelines using those artifacts, and confirmed that all jobs passed:
- End-to-end tests
- Nested end-to-end tests
- Connectivity tests
- ISA-95 smoke tests

## Azure IoT Edge PR checklist: