Use a different agent to build .NET (#7421)
We recently removed .NET 6.0 from all but one of our build agent images, because it is no longer supported. We kept it on agent-aziotedge-ubuntu-20.04-msmoby because we still have a dependency on .NET 6 for our Azure Functions module.

This change updates the agent we use when building .NET (6 and 8) to agent-aziotedge-ubuntu-20.04-msmoby, so that the build pipeline will continue to work following the agent updates described earlier.

I also changed other instances of agent-aziotedge-ubuntu-20.04-docker in the same file because Ubuntu 20.04 will soon reach end-of-life so this gives us a head start on migrating away from that version. But the bulk of the work to migrate away from Ubuntu 20.04 will go into a separate change.

To test, I ran the Build CI pipeline against these changes and confirmed that the .NET build passes.

## Azure IoT Edge PR checklist: