Add support for Azure Linux 3.0 (#7418)
This change adds support for building and testing IoT Edge for Azure Linux 3.0.

To test, I confirmed that the CI build passes, and that the right artifacts are produced. I also ran the end-to-end tests pipeline and confirmed that the Mariner and Azure Linux jobs (x64 and arm64) ran and passed.

## Azure IoT Edge PR checklist: