Fix iotedged checkin pipeline (#7426)
Within the last few days the itoedged checkin pipeline has started failing in the code coverage job. I'm not sure what changed with rust or cargo or the cargo-tarpaulin tool, but the tool is no longer happy with the environment. I was able to fix it by sourcing the cargo environment instead of manually setting `CARGO="${CARGO_HOME:-"$HOME/.cargo"}/bin/cargo"`.

## Azure IoT Edge PR checklist: