Remove use of rust-musl-builder image from API Proxy build.sh (#7390)
Resolves #7368 

Replaces use of ekidd/rust-musl-builder with cross-platform-rust-build.sh. This fixes errors that occur when build.sh runs (i.e. the OS Error 13 that occurs when attempting to create a tmp file to /opt/rust/rustup).

This change was tested on an ubuntu 20.04 dev machine by setting the ARCH env var to amd64 and running build.sh. This resulted in the azureiotedge-api-proxy image building successfully locally without any errors.

## Azure IoT Edge PR checklist: