Downgrade 1ES agents to Ubuntu 22.04 (#7435)
We recently upgraded several agents in our build pipelines from Ubuntu 20.04 to 24.04. It turns out that 1ES support for Ubuntu 24.04 is still in preview--I think this is why we're seeing lots of agent restarts. We're having to rerun pipelines several times to get a passing result.

This change downgrades agents to Ubuntu 22.04, which is supported by 1ES. In my testing, this eliminates the problem of agent restarts and makes our pipelines more reliable.

To test, I ran all the impacted pipelines and ensured they pass (checkin pipelines, CI build and test pipelines, Detect image updates, and Service deployment). I also ran the end-to-end tests against the bits from the CI build, and confirmed they pass.

## Azure IoT Edge PR checklist: