Only build components for the os/arch under test (#7429)
The end-to-end checkin pipeline combines logic from the CI build and end-to-end test pipelines to build and run end-to-end tests for a single OS/architecture (Ubuntu 22.04 amd64). However this pipeline does a lot of unnecessary work, e.g.:
- builds amd64 host packages for serveral operating systems
- spins up host package build agents for _all_ OS/architecture combinations, even though most of them don't do any work
- builds multi-architecture Docker images

This change updates the templates for the host package and Docker image build stages that are shared by the e2e-checkin and CI build pipelines. In the latter pipeline, all platforms/architectures will be built. In the former pipeline, only Ubuntu 22.04 amd64 will be built. This will prevent the needless spinning up of several build agents.

To test, I ran both the End-to-end Checkin and CI Build pipelines, and confirmed they pass. I also ran the End-to-end test pipeline against the artifacts from the CI Build run and confirmed all tests pass.

## Azure IoT Edge PR checklist: