Build API Proxy amd64 in an Alpine container (#7460)
The API Proxy module's amd64 image was recently updated to the latest stable version of Alpine, but it is still bringing in a statically linked build of OpenSSL v1.1.1i, which is out of support.

This change updates the module's amd64 build to _dynamically_ link to the musl and OpenSSL libraries provided by Alpine. This approach is simpler to build and lets us get OpenSSL updates automatically from the base image.

To test, I ran the CI pipeline and confirmed that it built API Proxy. Then I ran the ISA95 smoke test pipeline against the fresh build and confirmed that it passed.

## Azure IoT Edge PR checklist: