Use 1ES-hosted agent for Ubuntu 24.04 arm64 (#7465)
This change updates the Ubuntu 24.04 arm64 job in the end-to-end tests to use a 1ES-hosted agent instead of a custom agent. We created the corresponding 1ES-hosted image a while back but it didn't work because 1ES-internal software didn't support Ubuntu 24.04 arm64 at the time. That problem has since been resolved, so we can switch over now and decommision the custom agent VM we've been maintaining.

## Azure IoT Edge PR checklist: