Add arm64 architecture to ISA95 smoke tests (#7462)
This change updates the ISA95 smoke test pipeline to run tests on arm64 agents in addition to the existing support for amd64. This will give us coverage for the API Proxy module on arm64. Other pipelines (nested end-to-end, nested longhaul) which share some of the same pipeline logic, continue to run only on amd64.

I ran the ISA95 smoke test and nested end-to-end test pipelines and confirmed they passed and behaved as expected.

## Azure IoT Edge PR checklist: