Fix API Proxy module build to mark the entrypoint executable. (#7467)
The ADO release pipeline for this image ends up losing the +x mode bits on the binary because it uses ADO tasks that round-trip the binary through a .zip.

This went unnoticed in CI like the ISA95 pipeline, because those rely on images built by the regular build-images pipeline that does not have this problem, because it uses tasks that round-trip through a .tar instead.