connection test before report (#7414)
The main change is the call to `IsServiceAvailableViaTcpAsync` in the reconcile loop.

Now, it is possible to configure for an additional check before it reports upstream. The majority of the change is to propagate this settings and to figure out the host/port information, which it loads into a structure.

Note, that this is just a workaround for an sdk bug:
- the expected behavior is that the reporting fails with a network error.
- what happens, however, that the call keeps hanging indefinitely, blocking the reconciliation loop of edge agent. Although there is a timeout that interrupts this call, that restarts the reconciliation again, which then fails the same way as before, without doing any actual work.