Use 1ES-hosted agent for Debian 12 arm64 end-to-end tests (#7475)
Hosted Debian 12 arm64 images are now supported by 1ES. This change updates the end-to-end tests pipeline to use a 1ES hosted agent for the Debian12 arm64 job. It also removes a couple scripts related to the creation of custom agent VMs, as the scripts have not been maintained or used.

To test, I ran the end-to-end pipeline and verified that (1) the Debian 12 arm64 job used the 1ES-hosted agent, and (2) the Debian 12 arm64 job passed.

## Azure IoT Edge PR checklist: