Fix az CLI query in ISA-95 pipeline (#7479)
There's an Azure CLI command in the setup for the ISA-95 smoke test pipeline which gets the NSG for a given NIC. It needs this information so it can add a rule to the NSG to block outbound traffic during the smoke test. The JMESPath query in the Azure CLI command was wrong; it is supposed to find any NSGs associated with the NIC via network interface _or_ subnet, but it was only finding the former. This update fixes the query to find both.

I also replaced some hard-coded proxy URLs with pipeline variables. In the future we can replace those with Azure CLI commands that get the IP address for each proxy server given their VM name (which is currently still hard-coded into the tests), but this incremental improvement is a reasonable first step.

I ran the smoke test in a previously failing environment that has a subnet association instead of network interface, and it passed.

## Azure IoT Edge PR checklist: